refactor: rename different versions for better version distinction
