+++ examples/ApiExample.Modern/ModernSmsApiExample.csproj
