+++ examples/ApiExample.Modern/appsettings.json
