+++ examples/SendSmsExample.Modern/SendSmsExample.csproj
