+++ sdk/SMSApi.Legacy/SMSApi.pidb
