+++ sdk/SMSApi.Legacy/deps/Net35/Newtonsoft.Json.pdb
