+++ sdk/SMSApi.Legacy/deps/Net40/Newtonsoft.Json.pdb
