+++ sdk/SMSApi.Legacy/interfaces/IAccounts.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using InteractuaMovil.ContactoSms.Api;
+
+namespace InteractuaMovil.ContactoSms.Api.interfaces
+{
+    public interface IAccounts
+    {
+        ResponseObjects.ApiResponse<ResponseObjects.AccountStatusResponse> Status();
+    }
+}
