+++ sdk/SMSApi.Legacy/interfaces/IApiRequest.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace InteractuaMovil.ContactoSms.Api
+{
+    public interface IApiRequest
+    {
+
+        string BaseUrl
+        {
+            get;
+        }
+
+    }
+}
