--- sdk/SMSApi/interfaces/IAccounts.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using InteractuaMovil.ContactoSms.Api;
-
-namespace InteractuaMovil.ContactoSms.Api.interfaces
-{
-    public interface IAccounts
-    {
-        ResponseObjects.ApiResponse<ResponseObjects.AccountStatusResponse> Status();
-    }
-}
