--- sdk/SMSApi/interfaces/IApiRequest.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace InteractuaMovil.ContactoSms.Api
-{
-    public interface IApiRequest
-    {
-
-        string BaseUrl
-        {
-            get;
-        }
-
-    }
-}
